Refactoring de la selection org/pole vers un composable
