+++ web/src/pages/FollowUpCreationModificationPage.vue
@@ -9,7 +9,7 @@
 </route>
 
 <script setup lang="ts">
-import { computed, nextTick, ref, watch } from "vue"
+import { computed, nextTick, ref } from "vue"
 import SurveyBuilder from "../components/SurveyBuilder/index.vue"
 import { useApiFetch } from "../utils/data-fetching.ts"
 import { useToastStore } from "../stores/toast.ts"
@@ -22,6 +22,7 @@ import ProgressSpinner from "../components/ProgressSpinner.vue"
 import type { SurveySchema } from "@shared-types/survey"
 import IconPicker from "../components/IconPicker.vue"
 import EmptyPageTitleModal from "../components/SurveyBuilder/EmptyPageTitleModal.vue"
+import { useAdminScopeSelection } from "../composables/useAdminScopeSelection.ts"
 
 const store = useRootStore()
 const router = useRouter()
@@ -137,97 +138,17 @@ const isFetching = computed(
   () => isFetchingFollowUp.value || isFetchingSurvey.value
 )
 
-const adminMemberships = computed(() =>
-  store.loggedUser?.memberships.filter((x) => x.membershipType === "admin")
-)
-
-const uniqueAdminOrgs = computed(() => {
-  const seen = new Set<number>()
-  return (adminMemberships.value ?? [])
-    .filter((m) => {
-      if (seen.has(m.organisation.id)) return false
-      seen.add(m.organisation.id)
-      return true
-    })
-    .map((m) => m.organisation)
-})
-
-const orgOptions = computed(() =>
-  uniqueAdminOrgs.value.map((org) => ({
-    text: org.name,
-    value: String(org.id),
-  }))
-)
-
-const selectedOrganisationId = ref<string>("")
-
-const organisation = computed(() => {
-  if (uniqueAdminOrgs.value.length === 1) return uniqueAdminOrgs.value[0].id
-  return selectedOrganisationId.value
-    ? Number(selectedOrganisationId.value)
-    : undefined
-})
-
-// L'utilisateur a un rôle admin au niveau organisation (pole === null) pour l'org sélectionnée
-const hasOrgLevelAdmin = computed(
-  () =>
-    adminMemberships.value?.some(
-      (m) => m.organisation.id === organisation.value && m.pole === null
-    ) ?? false
-)
-
-// Pôles auxquels l'utilisateur a un rôle admin explicite dans l'org sélectionnée
-const adminPoles = computed(
-  () =>
-    adminMemberships.value
-      ?.filter(
-        (m) => m.organisation.id === organisation.value && m.pole !== null
-      )
-      .map((m) => m.pole!) ?? []
-)
-
-// Pôles de l'org sélectionnée issus du profil utilisateur (déjà chargés au login)
-const orgPoles = computed(
-  () =>
-    store.loggedUser?.organisations.find((o) => o.id === organisation.value)
-      ?.poles ?? []
-)
-
-// Toutes les valeurs sont des strings pour éviter les problèmes de coercition du DsfrSelect :
-// "" = aucun pôle, "123" = pôle avec id 123
-const poleOptions = computed(() => {
-  const opts: { text: string; value: string }[] = []
-  if (hasOrgLevelAdmin.value) {
-    // Les admins org voient "Aucun pôle" + tous les pôles de l'organisation
-    opts.push({ text: "Tous les pôles (niveau organisation)", value: "" })
-    for (const p of orgPoles.value) {
-      opts.push({ text: p.name, value: String(p.id) })
-    }
-  } else {
-    // Les admins pôle voient uniquement leurs pôles explicites
-    for (const p of adminPoles.value) {
-      opts.push({ text: p.name, value: String(p.id) })
-    }
-  }
-  return opts
-})
-
-// Le sélecteur n'est affiché que s'il y a un vrai choix à faire
-const showPoleSelect = computed(() => poleOptions.value.length > 1)
-
-const selectedPoleOption = ref<string>(
-  poleOptions.value.length === 1 ? poleOptions.value[0].value : ""
-)
-
-// Remise à zéro quand l'organisation change pour éviter une valeur obsolète
-watch(organisation, () => {
-  selectedPoleOption.value = ""
-})
-
-// null → enquête au niveau organisation ; number → enquête rattachée à un pôle
-const pole = computed(() =>
-  selectedPoleOption.value !== "" ? Number(selectedPoleOption.value) : null
-)
+const {
+  selectedOrganisationId,
+  selectedPoleOption,
+  organisation,
+  pole,
+  orgOptions,
+  poleOptions,
+  showPoleSelect,
+  hasOrgLevelAdmin,
+  uniqueAdminOrgs,
+} = useAdminScopeSelection()
 
 const createOrUpdateFollowUp = async () => {
   const hasPageWithoutTitle = schema.value.pages?.some(
